--- CharacterTool/EventMgr.cpp
@@ -1,1 +1,0 @@
-#include "EventMgr.h"
