--- CharacterTool/EventMgr.h
@@ -1,10 +1,0 @@
-#pragma once
-#include "Engine_include.h"
-
-class EventMgr
-{
-	SINGLETON(EventMgr)
-#define DATA EventMgr::GetInst()
-
-};
-
