+++ CharacterTool/src/main.cpp
@@ -1,4 +1,5 @@
 #include "CharacterTool.h"
+#include "Engine_include.h"
 
 int WINAPI wWinMain(HINSTANCE hinstance, HINSTANCE hPrevInstance, LPWSTR lpCmdLine, int nCmdShow)
 {
